utils/idir/requirements.txt
msgraph-sdk==1.4.0
azure-identity==1.17.1
